backend/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiohttp
cloudinary
supabase
passlib
python-jose
bcrypt
pydantic
python-dotenv
redis 
aioredis
